fix(arithmetic): match driver extern declarations to generated symbol types
## Summary

`arithmetic/driver.cpp` declares the globals produced by `kgen.py`-generated
kernel files as `extern int`, but the generated code defines them as
`unsigned long`. This is an ODR violation / type mismatch: reading the
values through the wrong type is undefined behavior and can report garbage
instruction/op counts on big-endian or strict-aliasing-sensitive builds.

## Root cause

Generated kernel files (see any `arithmetic/kernels/*.cpp`) define:

```cpp
unsigned long block_inst = 16;
unsigned long block_ops  = (16*(1*1));
unsigned long unroll     = 4;
```

while `driver.cpp` declared:

```cpp
extern int block_inst;
extern int block_ops;
extern int unroll;
```

Linking with `-flto -Wodr` confirms the mismatch:

```
warning: type of 'block_ops' does not match original declaration [-Wlto-type-mismatch]
warning: type of 'block_inst' does not match original declaration [-Wlto-type-mismatch]
warning: type of 'unroll' does not match original declaration [-Wlto-type-mismatch]
note: code may be misoptimized unless '-fno-strict-aliasing' is used
```

On little-endian aarch64 with small values this happens to read the correct
low 32 bits, which is why it went unnoticed; the values are still read
through the wrong type (UB).

## Fix

Change the three extern declarations in `driver.cpp` to `unsigned long` to
match the generated definitions. No other changes.

## Testing

Full kernel build/run requires an aarch64 (ARM) host — the kernels use ARM
NEON/SVE inline assembly — so it cannot run on this x86-64 machine.

Verified with a stub kernel object on x86-64 (g++ 13.3.0):

```
# Before fix:
g++ -O -flto -Wodr -o drv driver.cpp stub_kernel.cpp
# -> 3 x -Wlto-type-mismatch warnings (block_inst, block_ops, unroll)

# After fix:
g++ -O -flto -Wodr -o drv_fixed driver.cpp stub_kernel.cpp
# -> no warnings; program output identical (Inst/Iter=64, Ops/Iter=64)
```

## Why existing tests missed it

The project has no test suite; the only validation is building and running
the benchmarks on ARM hardware, where the mis-typed read happens to return
the correct low bytes on little-endian systems.

+++ arithmetic/driver.cpp
@@ -10,9 +10,9 @@
 #include <cstdlib>
 
 extern const char * description;
-extern int block_inst;
-extern int block_ops;
-extern int unroll;
+extern unsigned long block_inst;
+extern unsigned long block_ops;
+extern unsigned long unroll;
 extern void kernel(unsigned long);
 
 int main(int argc, char ** argv)
